# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API of tixkitz ticket systeme",
        "title": "LOKET TIKET",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6011",
    "basePath": "/",
    "paths": {
...
        "/cinemas/schedules": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all available cinema schedules from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cinemas"
                ],
                "summary": "Get cinema schedules",
                "responses": {
                    "200": {
                        "description": "Schedules fetched successfully or no schedules available",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error while fetching the schedule",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleResponse"
                        }
                    }
                }
            }
        },
        "/cinemas/{schedule_id}/seats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all available seats for a specific cinema schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cinemas"
                ],
                "summary": "Get available seats",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The ID of the cinema schedule",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available seats retrieved successfully or no seats available",
                        "schema": {
                            "$ref": "#/definitions/models.AvailSeatsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid schedule ID format",
                        "schema": {
                            "$ref": "#/definitions/models.AvailSeatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error while fetching available seats",
                        "schema": {
                            "$ref": "#/definitions/models.AvailSeatsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AvailSeatsResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Seat"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.CinemaSchedule": {
            "type": "object",
            "properties": {
                "cinema_name": {
                    "type": "string",
                    "example": "ebv"
                },
                "date": {
                    "type": "string"
                },
                "location": {
                    "type": "string",
                    "example": "Bogor"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Pulp Fiction"
                }
            }
        },
...
        "models.ScheduleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CinemaSchedule"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.Seat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 32
                },
                "pos": {
                    "type": "string",
                    "example": "C4"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}